feat: Receive PWM from Twist Controller of type Int16 and publish to CAN

+++ include/can_interface.hpp
@@ -9,7 +9,7 @@
 #include "rclcpp/rclcpp.hpp"
 #include "can_msgs/msg/frame.hpp"
 #include "std_msgs/msg/float64.hpp"
-#include "std_msgs/msg/float32.hpp"
+#include "std_msgs/msg/int16.hpp"
 #include "autoware_auto_vehicle_msgs/msg/velocity_report.hpp"
 #include "autoware_auto_vehicle_msgs/msg/steering_report.hpp"
 #include "autoware_auto_control_msgs/msg/ackermann_control_command.hpp"
@@ -37,7 +37,7 @@ class AwToCan : public rclcpp::Node
         rclcpp::Subscription<autoware_auto_control_msgs::msg::AckermannControlCommand>::SharedPtr sub_aw_command_;
         rclcpp::Subscription<std_msgs::msg::Float64>::SharedPtr TC_thro_cmd;
         rclcpp::Subscription<std_msgs::msg::Float64>::SharedPtr TC_brake_cmd;
-        rclcpp::Subscription<std_msgs::msg::Float32>::SharedPtr TC_steer_cmd;
+        rclcpp::Subscription<std_msgs::msg::Int16>::SharedPtr TC_steer_cmd;
 
         // CAN
         rclcpp::Subscription<can_msgs::msg::Frame>::SharedPtr sub_can_;
@@ -48,13 +48,13 @@ class AwToCan : public rclcpp::Node
         
         float TC_thro_output_cmd_ = 0.0;
         float TC_brake_output_cmd_ = 0.0;
-        float TC_steer_output_cmd_ = 0.0;
+        int16_t TC_steer_output_cmd_ = 0;
 
         void can_data_callback(const can_msgs::msg::Frame::SharedPtr msg);
         void AwCmd_callback(const autoware_auto_control_msgs::msg::AckermannControlCommand::SharedPtr msg);
         void TCthro_callback(const std_msgs::msg::Float64::SharedPtr msg);
         void TCbrake_callback(const std_msgs::msg::Float64::SharedPtr msg);
-        void TCsteer_callback(const std_msgs::msg::Float32::SharedPtr msg);
+        void TCsteer_callback(const std_msgs::msg::Int16::SharedPtr msg);
         void TimerCallback();
 };
